Limited messages for any request.  Fix NACK after ACK.
This will limit the total number of messages that can be sent
regarding a given request ID and protected against a type of
accidental DOS that we saw.  The problem had something to do
with a NACK after an ACK.

This patch also fixes some problems handling a NACK after an
ACK.  We were deleting the request_id information on NACK which
caused a KeyError when the command ultimately tried to reply
(which it must do).