Adding an extras package the agent archive
This patch adds an omnibus builder for an set of optional deps.  The
main one it includes right now is puppet.  This package can be installed
by the agent when a puppet command comes in.  This is better than using
a puppet distribution packages due to potentially conflicting ruby
versions etc.