Updates for the new virtualenv path
Agent version 0.11.0 had the virtualenv path changed.  For backward
compatibility a symlink was added to the old location.  However when
detecting if the agent was already installed we only look at the new
path.  Further when upgrading an agent with dpkg directories are left
in place and thus the symlink cannot be created.  This patch should fix
these problems.